Decoupled the sample generation and the unit tests. Streamer don't support application/oebps-package+xml, to be discussed
